move JavaExecutableFinder and JavaRuntimePropertiesDetector to util
