+++ openwebstart/src/test/java/com/openwebstart/jvm/util/JavaRuntimePropertiesDetectorTest.java
@@ -1,11 +1,10 @@
-package com.openwebstart.jvm.localfinder;
+package com.openwebstart.jvm.util;
 
-import com.openwebstart.jvm.localfinder.JavaRuntimePropertiesDetector.JavaRuntimeProperties;
+import com.openwebstart.jvm.util.JavaRuntimePropertiesDetector.JavaRuntimeProperties;
 import org.junit.jupiter.api.Test;
 
 import java.nio.file.FileSystems;
 import java.nio.file.Path;
-import java.nio.file.attribute.BasicFileAttributes;
 
 import static net.adoptopenjdk.icedteaweb.JvmPropertyConstants.JAVA_HOME;
 import static net.adoptopenjdk.icedteaweb.JvmPropertyConstants.JAVA_VENDOR;
@@ -28,10 +27,5 @@ class JavaRuntimePropertiesDetectorTest {
         assertEquals(System.getProperty(JAVA_VERSION), result.getVersion());
         assertEquals(System.getProperty(OS_NAME), result.getOsName());
         assertEquals(System.getProperty(OS_ARCH), result.getOsArch());
-    }
-
-    private boolean isJavaExecutable(Path path, BasicFileAttributes basicFileAttributes) {
-        final String fileName = path.getFileName().toString();
-        return fileName.equalsIgnoreCase("java") || fileName.equalsIgnoreCase("java.exe");
     }
 }
